# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultra ERP API Registry",
        "description": "Global modular API documentation for the Production-grade ERP system.",
        "contact": {
            "email": "admin@erp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Primary API Gateway"
        }
    ],
    "paths": {
...
        "/api/v1/refunds": {
            "get": {
                "tags": [
                    "Refunds"
                ],
                "summary": "List all refunds",
                "operationId": "bc37d267c878ed55e02d38be748244d6",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "sale_id": 5,
                                                "warehouse_id": 1,
                                                "status": "completed"
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Create new refund",
                "operationId": "cbf03d8335434a33279ca9349350c2d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_id",
                                    "warehouse_id",
                                    "items"
                                ],
                                "properties": {
                                    "sale_id": {
                                        "type": "integer"
                                    },
                                    "warehouse_id": {
                                        "type": "integer"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "sale_item_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Refund created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Refund created successfully.",
                                    "data": {
                                        "id": 2,
                                        "sale_id": 8,
                                        "warehouse_id": 2,
                                        "status": "pending"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/refunds/{refund}": {
            "get": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Get refund details",
                "operationId": "5a015544413a1d04461220fd7c4c26c2",
                "parameters": [
                    {
                        "name": "refund",
                        "in": "path",
                        "description": "Refund ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "sale_id": 5,
                                        "warehouse_id": 1,
                                        "notes": "Customer returned defective item",
                                        "status": "completed",
                                        "items": [
                                            {
                                                "id": 1,
                                                "sale_item_id": 12,
                                                "quantity": 1
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Refund not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Update refund status/notes",
                "operationId": "3ccdb617dd71ebf6e71edd2374b1db09",
                "parameters": [
                    {
                        "name": "refund",
                        "in": "path",
                        "description": "Refund ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refund updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Refund updated.",
                                    "data": {
                                        "id": 1,
                                        "status": "completed",
                                        "notes": "Refund processed"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Refund not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Delete refund record",
                "operationId": "3914507a56af35a628bf140ad992f55f",
                "parameters": [
                    {
                        "name": "refund",
                        "in": "path",
                        "description": "Refund ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Refund deleted successfully"
                    },
                    "404": {
                        "description": "Refund not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in 'Bearer {token}' format",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}